Scheduler, remove
Now adding scheduled commands returns a valid id which internally maps
to a pointer to the schedule node.  This id can be used to remove the
node from the schedule.
We are now deleting commands after their execute_done() method is
called.
the remove(ScheduleNode *) method now works because we set the mSchedule
pointer to node->next if node = mSchedule